Trust_or_lie/requirements.txt
python-telegram-bot
python-statemachine
